Move faction name pools into dedicated file
