+++ Assets/Scripts/Game/Factions/FactionNamePool.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 28bcacc55751417c8e142e3e3267bcc8
